Create Custom WordPress User Roles
Have you ever wanted a role which is higher than editor but not an administrator? Or maybe an author with extra benefits, but not quite an editor of the whole website?

This code snippet should only be run once and then commented out or deleted. It’s a code snippet which injects a new user role into your WordPress database, and defines all the unique privileges as well.