Create Force SSL HTTPS on Specific Post IDs
For checkout pages or unique signup/login pages you may wish to set HTTPS as the request header. This can be done using .htaccess but it’s also possible using a PHP code snippet. Be sure you’ve already created the page(s) and have an SSL license for the domain.